# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Foros, Mensajes y Respuestas",
        "description": "Documentación de la API para la gestión de foros, mensajes, respuestas y autenticación.",
        "contact": {
            "email": "1813370@alu.murciaeduca.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://catalogovideojuegos.test/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/foros": {
            "get": {
                "tags": [
                    "Foros"
                ],
                "summary": "Obtener lista de foros",
                "description": "Retorna una lista paginada de foros.",
                "operationId": "getForosList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de página para la paginación",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Foro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Foros"
                ],
                "summary": "Crear un nuevo foro",
                "description": "Crea un nuevo foro y retorna los datos del foro creado.",
                "operationId": "storeForo",
                "requestBody": {
                    "description": "Datos del foro a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForoApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Foro creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Foro"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/foros/{foro}": {
            "get": {
                "tags": [
                    "Foros"
                ],
                "summary": "Obtener información de un foro específico",
                "description": "Retorna los datos de un foro.",
                "operationId": "getForoById",
                "parameters": [
                    {
                        "name": "foro",
                        "in": "path",
                        "description": "ID del foro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Foro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Foro no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Foros"
                ],
                "summary": "Actualizar un foro existente",
                "description": "Actualiza los datos de un foro y retorna el foro actualizado.",
                "operationId": "updateForo",
                "parameters": [
                    {
                        "name": "foro",
                        "in": "path",
                        "description": "ID del foro a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del foro a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForoApiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foro actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Foro"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Foro no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Foros"
                ],
                "summary": "Eliminar un foro",
                "description": "Elimina un foro existente.",
                "operationId": "deleteForo",
                "parameters": [
                    {
                        "name": "foro",
                        "in": "path",
                        "description": "ID del foro a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foro eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Foro eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Foro eliminado exitosamente (sin contenido)"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Foro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserLite": {
                "title": "User Lite Information",
                "description": "Información básica del usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nombre Usuario"
                    }
                },
                "type": "object"
            },
            "VideojuegoConRol": {
                "title": "Videojuego con Rol",
                "description": "Información del videojuego asociado a un foro y su rol",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Nombre del Videojuego"
                    },
                    "imagen_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://example.com/image.jpg",
                        "nullable": true
                    },
                    "rol_en_foro": {
                        "type": "string",
                        "enum": [
                            "principal",
                            "secundario",
                            "opcional"
                        ],
                        "example": "principal"
                    }
                },
                "type": "object"
            },
            "RespuestaForo": {
                "title": "Respuesta de Mensaje de Foro",
                "description": "Representación de una respuesta a un mensaje de foro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "contenido": {
                        "type": "string",
                        "example": "Esta es una respuesta."
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/UserLite"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:35:00.000000Z"
                    }
                },
                "type": "object"
            },
            "MensajeForo": {
                "title": "Mensaje de Foro",
                "description": "Representación de un mensaje dentro de un foro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "contenido": {
                        "type": "string",
                        "example": "Contenido del mensaje principal."
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/UserLite"
                    },
                    "respuestas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RespuestaForo"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:05:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Foro": {
                "title": "Foro",
                "description": "Representación de un foro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Título del Foro"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción detallada del foro."
                    },
                    "creado_por": {
                        "$ref": "#/components/schemas/UserLite"
                    },
                    "videojuegos_asociados": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VideojuegoConRol"
                        }
                    },
                    "mensajes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MensajeForo"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T09:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T09:05:00.000000Z"
                    }
                },
                "type": "object"
            },
            "ErrorValidation": {
                "title": "Error de Validación",
                "description": "Objeto de error devuelto en caso de fallos de validación (422)",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Objeto con los errores de validación por campo",
                        "type": "object",
                        "example": {
                            "titulo": [
                                "El campo título es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "ForoApiRequest": {
                "title": "Foro API Request Body",
                "description": "Cuerpo de la petición para crear o actualizar un foro",
                "required": [
                    "titulo",
                    "descripcion"
                ],
                "properties": {
                    "titulo": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Nuevo Foro Sobre Estrategia"
                    },
                    "descripcion": {
                        "type": "string",
                        "maxLength": 5000,
                        "example": "Discusión sobre tácticas avanzadas."
                    },
                    "videojuegosConRoles": {
                        "description": "Objeto asociativo donde la clave es el ID del videojuego y el valor es el rol. Ejemplo: {'1': 'principal', '2': 'secundario'}",
                        "type": "object",
                        "example": {
                            "1": "principal",
                            "5": "secundario"
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "string",
                            "enum": [
                                "principal",
                                "secundario",
                                "opcional"
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Token Bearer (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}